+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 4.9.0 - (September 29, 2020)
 
 * Changed